refactor: split up auth modules
