+++ config/jest/jest.env.js
@@ -4,3 +4,11 @@ process.env.PUBLIC_API_URL = '/public-api'
 process.env.GOOGLE_TAG_MANAGER_ID = 'GTM-123'
 process.env.UNLEASH_SERVER_API_URL = 'http://fakedfeatureflag.gov.uk'
 process.env.UNLEASH_SERVER_API_TOKEN = '123'
+
+process.env.AUTH_ENABLED = false
+process.env.NEXTAUTH_URL = 'http://fake-frontend.gov.uk'
+process.env.AUTH_SECRET = 123
+process.env.AUTH_DOMAIN = 'https://fake-env-domain.auth.eu-west-2.amazoncognito.com'
+process.env.AUTH_CLIENT_ID = 456
+process.env.AUTH_CLIENT_SECRET = 789
+process.env.AUTH_CLIENT_URL = 'https://cognito-idp.eu-west-2.amazonaws.com/eu-west-2_fake_env'
