+++ src/api/requests/helpers.spec.ts
@@ -1,4 +1,4 @@
-import { getApiBaseUrl, getCmsApiPath, requestOptions } from './helpers'
+import { getApiBaseUrl, getAuthApiBaseUrl, getCmsApiPath, requestOptions } from './helpers'
 
 /**
  * The process.env mocks powering these tests live inside <rootDir>/config/jest/jest.env.js
@@ -9,6 +9,10 @@ test('Get the API url', () => {
 
 test('Get the CMS API path', () => {
   expect(getCmsApiPath()).toEqual('http://fake-backend.gov.uk/api/pages')
+})
+
+test('Get the Cognito OAuth API path', () => {
+  expect(getAuthApiBaseUrl()).toEqual('https://fake-env-domain.auth.eu-west-2.amazoncognito.com/oauth2')
 })
 
 test('Fetch request options', () => {
